requirements.txt
alembic==1.8.1
aniso8601==9.0.1
arrow==1.2.3
attrs==22.2.0
autoenv==1.0.0
blinker==1.4
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
cryptography==38.0.1
dnspython==2.2.1
docopt==0.6.2
dominate==2.7.0
et-xmlfile==1.1.0
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-Login==0.6.2
Flask-Migrate==3.1.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-Uploads==0.2.1
Flask-WTF==1.0.1
google-api-core==2.11.0
google-auth==2.15.0
google-cloud-core==2.3.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
numpy==1.24.2
openpyxl==3.1.1
packaging==22.0
pandas==1.5.3
passlib==1.7.4
pipreqs==0.4.11
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.15.0
PyMySQL==1.0.2
PyPDF2==2.11.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-editor==1.0.4
python-magic==0.4.27
python-magic-bin==0.4.14  #comment first, then install this
pytz==2022.7
PyYAML==6.0
requests==2.28.1
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.41
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.1.2
WTForms==3.0.1
XlsxWriter==3.0.9
yarg==0.1.9